Create a different component for the search box instead of having it integrated in the NavBar component
