+++ src/routes/Home.jsx
@@ -1,6 +1,7 @@
 import NearestCity from 'components/NearestCity';
 import OtherCitiesList from 'components/OtherCitiesList';
 import SelectedCity from 'components/SelectedCity';
+import SearchBar from 'components/SearchBar';
 import { getData } from 'redux/home/homeSlice';
 import { useSelector } from 'react-redux';
 
@@ -10,6 +11,7 @@ const Home = () => {
   } = useSelector(getData);
   return (
     <div>
+      <SearchBar />
       {selectedCity && Object.keys(selectedCity).length !== 0 ? <SelectedCity />
         : <NearestCity city={city} isLoading={isLoading} error={error} /> }
       <OtherCitiesList otherCities={otherCities} isLoading={isLoading} error={error} />
