+++ src/styles/navbar.scss
@@ -12,46 +12,6 @@
       width: 40%;
     }
 
-    div.search-box {
-      display: flex;
-      align-items: center;
-
-      input {
-        height: 30px;
-        padding: 5px;
-        border-radius: 5px;
-        width: 100%;
-      }
-
-      ul.options-list {
-        position: absolute;
-        top: 50px;
-        width: 40%;
-        box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.3);
-        background-color: #f1f6f9;
-        padding-left: 5%;
-        font-weight: bold;
-        display: flex;
-        flex-direction: column;
-        gap: 10px;
-        font-size: 0.9rem;
-
-        a {
-          color: #212a3e;
-        }
-      }
-
-      .search-icon {
-        position: absolute;
-        right: 18%;
-        color: #212a3e;
-      }
-    }
-
-    input::placeholder {
-      padding: 5px;
-    }
-
     .home-icon {
       color: #f1f6f9;
       width: 25px;
